+++ CHANGELOG.md
@@ -16,6 +16,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.1] - 2018-04-20
+
+### Changed
+
+- Updated test suite to use molecule
+
+### Removed
+
+- Removed tests for Debian Jessie
+
 ## [1.2.0] - 2017-11-29
 
 ### Changed
@@ -28,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Ansible Galaxy metadata file.
 
-[Unreleased]: https://github.com/joshuacherry/ansible-role-openssl/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/joshuacherry/ansible-role-openssl/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/joshuacherry/ansible-role-openssl/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/joshuacherry/ansible-role-openssl/compare/1.0.1...1.2.0
 [1.0.1]: https://github.com/joshuacherry/ansible-role-openssl/compare/1.0.0...1.0.1